feature(core): Improve lawn and mower tests by using ScalaTest Table feature.

+++ lawn-mower/lawn-mower-core/src/main/scala/fr/mistertie/lawnmower/core/lawn/Lawn.scala
@@ -111,5 +111,5 @@ object Lawn {
    * @param topRightOrdinate the top right point's ordinate.
    * @return the created instance.
    */
-  def apply(topRightAbscissa: Short, topRightOrdinate: Short) = new Lawn(topRightAbscissa, topRightOrdinate)
+  def apply(topRightAbscissa: Int, topRightOrdinate: Int) = new Lawn(topRightAbscissa, topRightOrdinate)
 }
